requirements.txt
aiogram==3.15.0
APScheduler==3.11.0
python-dotenv==1.0.1
pytz==2024.2
Requests==2.32.3
moviepy==2.1.1